Make go-spacemesh linter happy

+++ examples/ex2_scale.go
@@ -9,12 +9,12 @@ import (
 func (t *Ex2) EncodeScale(enc *scale.Encoder) (total int, err error) {
 	if n, err := scale.EncodeStructSlice(enc, t.Slice); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
 	}
 	if n, err := scale.EncodeStructArray(enc, t.Array[:]); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
 	}
 	return total, nil
@@ -23,22 +23,22 @@ func (t *Ex2) EncodeScale(enc *scale.Encoder) (total int, err error) {
 func (t *Ex2) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if field, n, err := scale.DecodeStructSlice[Ex2](dec); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
 		t.Slice = field
 	}
 	if n, err := scale.DecodeStructArray(dec, t.Array[:]); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
 	}
 	return total, nil
 }
 
 func (t *Smth) EncodeScale(enc *scale.Encoder) (total int, err error) {
-	if n, err := scale.EncodeCompact32(enc, t.Val); err != nil {
+	if n, err := scale.EncodeCompact32(enc, uint32(t.Val)); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
 	}
 	return total, nil
@@ -47,9 +47,9 @@ func (t *Smth) EncodeScale(enc *scale.Encoder) (total int, err error) {
 func (t *Smth) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if field, n, err := scale.DecodeCompact32(dec); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
-		t.Val = field
+		t.Val = uint32(field)
 	}
 	return total, nil
 }
@@ -57,7 +57,7 @@ func (t *Smth) DecodeScale(dec *scale.Decoder) (total int, err error) {
 func (t *StructSliceWithLimit) EncodeScale(enc *scale.Encoder) (total int, err error) {
 	if n, err := scale.EncodeStructSliceWithLimit(enc, t.Slice, 2); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
 	}
 	return total, nil
@@ -66,7 +66,7 @@ func (t *StructSliceWithLimit) EncodeScale(enc *scale.Encoder) (total int, err e
 func (t *StructSliceWithLimit) DecodeScale(dec *scale.Decoder) (total int, err error) {
 	if field, n, err := scale.DecodeStructSliceWithLimit[Smth](dec, 2); err != nil {
 		return total, err
-	} else {
+	} else { // nolint
 		total += n
 		t.Slice = field
 	}
